moved utility classes to own folder
