+++ backend_vars.py
@@ -1,7 +1,7 @@
-import logger
-from config import threadConfiguration
-from database import threadDatabase
-from feed import NewsAPICalls
+import utils.logger as logger
+from utils.config import threadConfiguration
+from utils.database import threadDatabase
+from utils.feed import NewsAPICalls
 from flask_apscheduler import APScheduler
 
 log = logger.setup_logger('root')
